tcc: Move new parser package to new subdir
