+++ tcc/tcc/ir/generator.hpp
@@ -10,8 +10,8 @@
 #include <vector>
 
 #include "tcc/ir/intermediate_representation.hpp"
-#include "tcc/parser-qi/ast.hpp"
-#include "tcc/parser-qi/error_handler.hpp"
+#include "tcc/parser/ast.hpp"
+#include "tcc/parser/error_handler.hpp"
 
 namespace tcc {
