+++ tcc/tcc/parser/compiler.hpp
@@ -9,8 +9,8 @@
 #include <map>
 #include <vector>
 
-#include "tcc/parser-qi/ast.hpp"
-#include "tcc/parser-qi/error_handler.hpp"
+#include "tcc/parser/ast.hpp"
+#include "tcc/parser/error_handler.hpp"
 
 namespace tcc {
 namespace code_gen {
