+++ tcc/tcc/parser/expression.cpp
@@ -2,7 +2,7 @@
 #pragma warning(disable : 4345)
 #endif
 
-#include "tcc/parser-qi/statement_def.hpp"
+#include "tcc/parser/expression_def.hpp"
 
 using iterator_type = std::string::const_iterator;
-template struct tcc::parser::Statement<iterator_type>;
+template struct tcc::parser::Expression<iterator_type>;
