+++ tcc/tcc/parser/expression_def.hpp
@@ -1,8 +1,8 @@
 #include <boost/spirit/include/phoenix_function.hpp>
 
-#include "tcc/parser-qi/annotation.hpp"
-#include "tcc/parser-qi/error_handler.hpp"
-#include "tcc/parser-qi/expression.hpp"
+#include "tcc/parser/annotation.hpp"
+#include "tcc/parser/error_handler.hpp"
+#include "tcc/parser/expression.hpp"
 
 namespace tcc {
 namespace parser {
