+++ tcc/tcc/parser/function.cpp
@@ -2,7 +2,7 @@
 #pragma warning(disable : 4345)
 #endif
 
-#include "tcc/parser-qi/function_def.hpp"
+#include "tcc/parser/function_def.hpp"
 
 using iterator_type = std::string::const_iterator;
 template struct tcc::parser::function<iterator_type>;
