+++ tcc/tcc/parser/statement_def.hpp
@@ -1,6 +1,6 @@
-#include "tcc/parser-qi/annotation.hpp"
-#include "tcc/parser-qi/error_handler.hpp"
-#include "tcc/parser-qi/statement.hpp"
+#include "tcc/parser/annotation.hpp"
+#include "tcc/parser/error_handler.hpp"
+#include "tcc/parser/statement.hpp"
 
 namespace tcc {
 namespace parser {
