+++ tcc/tcc/parser/statement_test.cpp
@@ -3,14 +3,14 @@
  * @copyright Copyright 2019-2020 Tobias Hienzsch. MIT license.
  */
 
-#include "tcc/parser-qi/statement.hpp"
+#include "tcc/parser/statement.hpp"
 
 #include <sstream>
 
 #include "catch2/catch.hpp"
-#include "tcc/parser-qi/skipper.hpp"
+#include "tcc/parser/skipper.hpp"
 
-TEST_CASE("tcc/parser-qi: StatementValid", "[tcc][parser][qi]") {
+TEST_CASE("tcc/parser: StatementValid", "[tcc][parser][qi]") {
   using namespace tcc;
 
   auto testCase = GENERATE(as<std::string>{}, "int x = 0;", "x = y+z;");
@@ -29,7 +29,7 @@ TEST_CASE("tcc/parser-qi: StatementValid", "[tcc][parser][qi]") {
   REQUIRE(success == true);
 }
 
-TEST_CASE("tcc/parser-qi: StatementInvalid", "[tcc][parser][qi]") {
+TEST_CASE("tcc/parser: StatementInvalid", "[tcc][parser][qi]") {
   using namespace tcc;
 
   auto testCase = GENERATE(as<std::string>{}, "auto x = 0;", "= y+z;");
@@ -47,7 +47,7 @@ TEST_CASE("tcc/parser-qi: StatementInvalid", "[tcc][parser][qi]") {
 
   REQUIRE(success == false);
 }
-TEST_CASE("tcc/parser-qi: ReturnStatementValid", "[tcc][parser][qi]") {
+TEST_CASE("tcc/parser: ReturnStatementValid", "[tcc][parser][qi]") {
   using namespace tcc;
 
   auto testCase = GENERATE(as<std::string>{}, "return;", "return x;", "return 1+2;");
